Install illuminate/view and override constructor, remove custom implementation
