+++ composer.json
@@ -14,7 +14,8 @@
         "php": ">=7.3",
         "nikic/php-parser": "^4.5",
         "symfony/console": "^4.3 || ^5.0",
-        "symfony/process": "^4.3 || ^5.0"
+        "symfony/process": "^4.3 || ^5.0",
+        "illuminate/view": "^7.22.0 || ^8.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.3"
