--- src/Illuminate/Compilers/Concerns/CompilesHelpers.php
@@ -1,49 +1,0 @@
-<?php
-
-namespace Tighten\Illuminate\Compilers\Concerns;
-
-trait CompilesHelpers
-{
-    /**
-     * Compile the CSRF statements into valid PHP.
-     *
-     * @return string
-     */
-    protected function compileCsrf()
-    {
-        return '<?php echo csrf_field(); ?>';
-    }
-
-    /**
-     * Compile the "dd" statements into valid PHP.
-     *
-     * @param  string  $arguments
-     * @return string
-     */
-    protected function compileDd($arguments)
-    {
-        return "<?php dd{$arguments}; ?>";
-    }
-
-    /**
-     * Compile the "dump" statements into valid PHP.
-     *
-     * @param  string  $arguments
-     * @return string
-     */
-    protected function compileDump($arguments)
-    {
-        return "<?php dump{$arguments}; ?>";
-    }
-
-    /**
-     * Compile the method statements into valid PHP.
-     *
-     * @param  string  $method
-     * @return string
-     */
-    protected function compileMethod($method)
-    {
-        return "<?php echo method_field{$method}; ?>";
-    }
-}
