--- src/Illuminate/Compilers/Concerns/CompilesRawPhp.php
@@ -1,32 +1,0 @@
-<?php
-
-namespace Tighten\Illuminate\Compilers\Concerns;
-
-trait CompilesRawPhp
-{
-    /**
-     * Compile the raw PHP statements into valid PHP.
-     *
-     * @param  string  $expression
-     * @return string
-     */
-    protected function compilePhp($expression)
-    {
-        if ($expression) {
-            return "<?php {$expression}; ?>";
-        }
-
-        return '@php';
-    }
-
-    /**
-     * Compile the unset statements into valid PHP.
-     *
-     * @param  string  $expression
-     * @return string
-     */
-    protected function compileUnset($expression)
-    {
-        return "<?php unset{$expression}; ?>";
-    }
-}
